message-route: Ajout du controller  - MessageController: Création du controller pour /api/messages  - création des DTO MessageResponse et MessageRequest  - construction de MessageResponse dans MessageService  - UserManagementService: modification du type de retour pour getUserId() de Integer vers Long et propagation des changements  - Fix UserProfileResponse : ajout de l'id et propagation des changements pour renvoyer l'id suite à requète /me

+++ src/main/java/fr/jerem/chaotop_backend/controller/RentalContoller.java
@@ -111,13 +111,13 @@ public class RentalContoller {
         }
 
         String email = optionalAuthenticatedUserEmail.get();
-        Optional<Integer> optionalUserId = userManagementService.getUserId(email);
+        Optional<Long> optionalUserId = userManagementService.getUserId(email);
         if (optionalUserId.isEmpty()) {
             log.error("No Id found for email {}", email);
             return ResponseEntity.internalServerError().build();
         }
 
-        Integer userId = optionalUserId.get();
+        Long userId = optionalUserId.get();
         try {
             // TODO a specific service to store image
